Refactor to use EtherscanURIBackend
