+++ .circleci/config.yml
@@ -66,7 +66,7 @@ jobs:
         environment:
           TOXENV: py37-core
           WEB3_INFURA_PROJECT_ID: 4f1a358967c7474aae6f8f4a7698aefc
-          ETHPM_CLI_ETHERSCAN_API_KEY: RYKF6YJS1N2F5PUFA9Q1P9VG92R8N7NE3G 
+          ETHPM_CLI_ETHERSCAN_API_KEY: RYKF6YJS1N2F5PUFA9Q1P9VG92R8N7NE3G
 
 workflows:
   version: 2
